requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.3.0
contourpy==1.3.1
cycler==0.12.1
Flask==3.0.0
fonttools==4.56.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.10.0
mdurl==0.1.2
numpy==2.2.3
-e git+ssh://git@github.com/ordotools/ordotools.git@c197f21bb3499c0e0202f2f594d9fab5b96711f8#egg=ordotools
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
PyYAML==6.0.3
rich==13.9.4
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.3
